added security to views

+++ services/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Service, Category
 from django.db.models.functions import Lower
@@ -71,8 +72,13 @@ def service_detail(request, service_id):
     return render(request, 'services/service_detail.html', context)
 
 
+@login_required
 def add_service(request):
     """ Add a service to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ServiceForm(request.POST, request.FILES)
         if form.is_valid():
@@ -92,8 +98,13 @@ def add_service(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_service(request, service_id):
     """ Edit a service in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     service = get_object_or_404(Service, pk=service_id)
     if request.method == 'POST':
         form = ServiceForm(request.POST, request.FILES, instance=service)
@@ -116,10 +127,14 @@ def edit_service(request, service_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_service(request, service_id):
     """ Delete a service from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     service = get_object_or_404(Service, pk=service_id)
     service.delete()
     messages.success(request, 'Service deleted!')
-    return redirect(reverse('services')) 
-
+    return redirect(reverse('services'))
